requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
websockets==12.0
pydantic==2.8.2
openai==1.40.6
aiofiles==24.1.0
numpy==2.0.1
pydub==0.25.1
soundfile==0.12.1
# Optional ASR backends:
faster-whisper==1.0.3
# Optional TTS
pyttsx3==2.90
# RAG:
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.3.1
python-multipart==0.0.9
